Completed task 6

+++ 0x00-hello_world/6-size.c
@@ -7,10 +7,10 @@
  */
 int main(void)
 {
-printf("Size of a char: %d \n", sizeof(char));
-printf("Size of an int: %d \n", sizeof(int));
-printf("Size of a long int: %d \n", sizeof(long));
-printf("Size of a long long int: %d \n", sizeof(long int));
-printf("Size of a float: %d \n", sizeof(float));
+printf("Size of a char: %d \n", (int)sizeof(char));
+printf("Size of an int: %d \n", (int)sizeof(int));
+printf("Size of a long int: %d \n", (int)sizeof(long));
+printf("Size of a long long int: %d \n", (int)sizeof(long int));
+printf("Size of a float: %d \n", (int)sizeof(float));
 return (0);
 }
